NMS-15461: move to the correct location, provide a shim for backwards-compatibility
